Disable shuffled testing for tests with OutputMatching
Tests that use the OutputMatching plugin do not support shuffled
testing. In the past run_test handled this by reacting to the
TEST_EXPECTED environment variable. This environment variable was
deprecated when tap-wrapper was moved to a plugin-based structure.
This change adjusts the condition in run_test when to disable shuffled
testing.

Fixes #CD-864

Change-Id: Id3ecb34326a108c7eab98c9de865890db0a2158e