tap-wrapper: Add plugin features
Some plugins may not support all operations the test framework
supplies. This change introduces a feature flag where each plugin can
indicate if it supports a particular feature. The first user is test
shuffling.

Tests that use the OutputMatching plugin do not support shuffled
testing. In the past run_test handled this by reacting to the
TEST_EXPECTED environment variable. This environment variable was
deprecated when tap-wrapper was moved to a plugin-based structure.
This change moves the check for shuffling support to tap-wrapper and
disables shuffling when any plugin does not support it.

Change-Id: Idd4ff0fcba836c5e44beb8b3487f0d691b32c024