tap-wrapper: Add support for output filters
Some functionality of the system may require that output is
pre-processed before it can be passed to tap-wrapper plugins. This
change introduces the concept of filters that take a single line and
based on that line produce any number of output lines. Filters are
expected to be chained and the order they are specified in matters.

Change-Id: I46429aa48d070d2549613454a8beaf011d32892e