tap-wrapper: don't remove carriage return
Some plugins may rely on the whole unprocessed output being available to
them in order to produce correct results. If any consumer of data
requires processing of line endings it must do so on its own or the
feature must be integrated into the framework in another place, similar
to escape sequence removal.

Change-Id: I47886bc6eae40ef6ab8b1359e46bc8a29f889d3c