genexportpack: Always copy files unless copied for another entry
Up until now if files already existed they were not copied. This is
because when combining multiple module list entries into one
exportpack, files that were used by several entries are only copied
once into the exportpack. This heuristic however breaks when files
that are supposed to be copied already exist before genexportpack is
started. To distinguish between files that already existed before
genexportpack started and files that were already copied during the
current execution of genexportpack, we create a list of files created
and check that instead of the existance of the file.

Change-Id: I569d1b56d6aad4d704f4dbd4b9d1bdd600211cb7