mk: enable building with clang
With this commit one can now specify `make CLANG=...` to build using
LLVM/CLANG. There are three possible arguments and they are modelled
after what Linux does.
  1. If the argument starts with `-`, it is interpreted as a version
     number. E.g., `make CLANG=-14`. Then `clang-14`, `clang++-14`, and
     `clang-cpp-14` are expected to be found in `$PATH` and are used for
     building.
  2. If the arguments end in a slash, it is interpreted as a path. E.g.,
     `make CLANG=/usr/local/bin/`. Then `clang`, `clang++`, and
     `clang-cpp` are expected to be found in this directory.
  3. If it is anything else, conventionally `make CLANG=1`, `clang`,
     `clang++`, and `clang-cpp` are expected to be found in `$PATH`.

`CLANG=...` can also be added to `Makeconf.local`.

Building with clang thus only switches out the C and C++ compiler and
the preprocessor. For everything else, in particular the linker, the GNU
toolchain is still used.

Note that Clang uses binaries from an installed gcc toolchain as well:
See 'clang -v' and look for 'Selected GCC installation'. In particular
this is important for
 - GCCLIB_{HOST,EH,S_SO}, and
 - GCCLIB_FILE_{crtbegin.o,crtbeginS.o,crtbeginT,crtendS,crtend.o}

To first build with clang in an existing build directory, regenerate the
variables in `.config.all`. I.e., using option 3 from above, one needs
to do
  1. Create `$BUILDDIR/Makeconf.local` containing `CLANG=1`
  1. `make O=$BUILDDIR oldconfig`
  2. `make O=$BUILDDIR`

For a new builddir, the following works:
  1. `make B=$BUILDDIR`
  2. create `$BUILDDIR/Makeconf.local` containing `CLANG=1`
  3. `make O=$BUILDDIR`

Change-Id: I0c18a6d122d36a0b0ad09d56123fffac8b48bd88
Requires-Change-Id: I3e88309b5cc5f00321ef3a4e2270ef4f629b8fea