mk: build assember files without executable stack
We enforce that the stack is not executable at link time. When linking
outside of BID this is not necessarily the case. The compiler will add
the '.note.GNU-stack' section itself but the assembler must be
explicitly instructed to do so. Without the switch the linker will issue
a warning about the missing special 'GNU-stack' note section.

Change-Id: I4bb5d74f6ea537e5bcbece57af6ac993caf721cd