tool/bin/run_test - Use TEST_FIASCOCONFIG for KERNEL_CONF check
Up until now the path of the kernel config for KERNEL_CONF checks was
determined by extracting the dirname from the KERNEL environment
variable. This fails whenever the KERNEL environment variable is not
specified by the user as the default value is simply 'fiasco'. In this
case the kernel is found later using the MODULE_SEARCH_PATH, which is
too late for any KERNEL_CONF checks. In an effort to improve this we
use the TEST_FIASCOCONFIG environment variable first before resorting
to the KERNEL variable. The former is used for the test_env.lua file
already.

Change-Id: Id127616d510c7d60de1c5762d18b35ccc37e3627