bid-bender: libgcc after libgcc_eh
At least on ARM64 with gcc-13, libgcc_eh.a depends on a function
__aarch64_cas8_acq_rel() which is provided by libgcc.a. Hence, during
link, add libgcc.a after libgcc_eh.a to resolve the open reference.

Without this fix, at least libsupc++.so will contain an undefined
reference to this function and linking against libsupc++.so will fail
because __aarch64_cas8_acq_rel is a hidden symbol in libgcc.a.

Change-Id: Ib10724beec88e6412fb95c9fd384d40e5c78e605