Automatically rebuild RELOC_PHYS targets when changing RAM_BASE
Up until now switching the RAM_BASE only triggered a remake in
selected components, leaving out other components that may or may not
be present. This change introduces a few lines of code that
automatically register paths with the prog.mk role that have
RELOC_PHYS set to y in a dedicated file in the root of the build
directory. This file can then be read by the root Makefile to trigger
make again in these directories when the RAM_BASE changes.

Change-Id: I0bf2283c5df10a9638076e657d7f2b33a57a1c43