OutputMatching.pm - Add a fixed tap description
To make the ok not ok lines more distinguishable we introduce a
description parameter, which can take a TAP description like
Atkins::FooBar, which is used in the ok/not ok line instead of the
message resulting for the output matching. If the description is
present we use it and print the output matching message as a comment
below. If the description is not present we use the output matching
message as test description, like we did before.

Change-Id: Iefb37707c6ccf3976dd1ca6569661ee6b28e5362