mk: Add assets mode
Add an 'assets' mode that shall be used
for installing non programming / binary
type of files, such as device-tree blobs.

Previously we built device-tree blobs using the lib role
which meant that the dtb-files were installed into
lib/arch/l4f/*.dtb. This is misleading as a dtb-file is
not bound to an architecture or L4 variant.

With this change, dtb files are installed
into assets/dtb of the objdir and the assets
directory is added to the default search paths.

Change-Id: Id3baebac6aed2e91777337bb8253506829145fd9