l4image/ModList: Move module type handling to ModList.pm
In l4image the types of modules are part of the module info structure.
Currently the "magic" indices from ModList.pm are translated by l4image
to the types. Move this type processing to ModList.pm, adding it as a
hash member for the Module hash and use that in l4image.

This is in preparation for allowing to omit the specification of root
task or root pager, which is useful for kernel unit tests and for
testing of bootstrap.

Change-Id: If3a91b18debb1b732ed5cb0207a60f868a948f40