arm64: Add option to configure W^X for roottask
This option allows to enable W^X support for the roottask of the system.
Since there are drawbacks to this with only very limited usability
beyond some very specific use-cases the option defaults to off.

Having the option disabled allows the root task to self-patch to the
correct execution mode (syscall instruction). Having it enabled will
cause the system to stop working when moe tries to patch the
instruction. This only causes an issue in cases where the userland was
misconfigured, where the system would not work eitherway because it
could not call to the kernel.

Add a detailed help text to the option to document the limitations of
the W^X option.

Change-Id: I2010f37aab063cf4d021718561d11a5a6f714b56