run_test - Support options with spaces within check_config
To be able to check for the lack of support using the KERNEL_CONF
variable and others that use the same mechanism, a test developer would
have to specify a line such as "# CONFIG_JDB is not set". Since
check_config splits multiple options by space this is not easily
possible. To add support for this we replace the split method with
something similar that honors double-quotes while splitting, so that
such a line can be specified as

KERNEL_CONF = ". CONFIG_JDB is not set"

We use a . here instead of # because make interprets the latter as
comment. Further more we adjust test.mk so it escapes double-quotes in
variables.

Change-Id: I736861209e123ae27dfc2c0b4ad7186f355df10c