Makefile: unify version printing in report
This unifies the often repeating and very similar version output
commands. The change mostly preserves the output structure of the
previous version except that it sometimes does insert a newline after
the name of the printed command where there was none before.

In the interest of more unified output this can be accepted.

Change-Id: I2349ba5016d1ff218cd3f60960c107f80fc0f752