tool/imagebuilder/ - Fix post image cmd
The POST_IMAGE_CMD implementation was a bit wonky. Passing make code via
stdin didn't seem to work, so we're using a pipe now. Furthermore we set
the goal running the POST_IMAGE_CMD to phony now so make does not skip
it. We also don't need to set the POST_IMAGE_CMD explicitly as using the
call function on an environment variable is perfectly fine.

Change-Id: I3dfc4ca6b6fd9679fb165b76219e1db6feef4bf1