Grub.pm: Add support for creating hybrid ISO/UEFI images
The grub-mkrescue tool actually allows for both PC and UEFI modules to
be installed simultaneously and creates a hybrid PC/UEFI ISO image in
that case.

If we detect that grub-mkrescue neither creates an exclusive PC ISO nor
an exclusive UEFI ISO, enable the hybrid ISO detection as the last
resort.

Change-Id: I90bd16f1dbbebd1a8c80822eca6f4198fff3b69c