pkgdeps: ensure deep nested packages are built
For packages which have a deeper directory structure in a package style,
such as pkg/io, pkg/l4re-core etc. there is currently no dependency
generated such that pkg/io is built when invoking `make pkg/`.

Generate the fitting dependencies as well, such that any cluster that
has a name which is a path creates dependencies for the upwards
components of the path.

Change-Id: I66c723b8c18337cd62ba54e9e8442d479a3bfa1b