mk: ensure removed platforms don't remain in list
The platform lists and dependent include files for the build system must
depend on the platform definitions that were there on generation so we
can detect if one vanishes. So create the dep file accordingly and
ensure that _only_ the things coming from the Makefile (new platforms)
not old ones from the dependency file are written to the new dep file.

This is ensured by explicitly using the new list of platforms as
dependency and not $^ as it would also contain the dependencies from the
deps file.

Change-Id: Ia050a999542616fe973d964e20a0b5e5fc4fb8d6