Makefile - Enable make test to only run kunit-tests
Up until now it is only possible to run bid-tests and kunit-tests or
bid-tests alone, but not kunit-tests alone. To make it possible for a
user to run only kunit-tests we make the TESTS variable more powerful by
just passing it directly to prove.

To keep the old default behaviour TESTS defaults to either "bid-tests/"
or "bid-tests/ kunit-tests/" depending on whether or not TEST_KUNIT_DIR
is set. For backwards compatibility reasons, in particular that
specifying TESTS without the bid-tests/ prefix, we prepend bid-tests/ to
all paths in TESTS that neither have bid-tests/ nor kunit-tests/ as a
prefix.

Another detail of setting TESTS explicitly is that if the TESTS
environment variable is set outside of the L4Re build system possibly
for a different purpose, since it's such a commonly used name, it will
not implicitly influence the make test target. Instead make test
TESTS=$TESTS can be used for case where this is desired.

Change-Id: Id13d1859e732d3e2836acbcbdba275411f7bff9a