introspection tests: replace DEBUGPRINT action; rework output
For debug output, the sandbox environment now contains a print()
function. This function formats any arguments as TAP comments. Also
inspect() is handed into the sandbox. This replaces the DEBUGPRINT
action, which is removed.

The output is reworked such that TAP comment output is only generated
when an action fails or the sandbox uses print(). In such a case, also
the filename of the responsible snippet is printed.

Change-Id: I37975dc00541a379514c50a70741bb3f8594246f