l4image - Add edit mode for launch command
This change adds a quick way to edit the command line used by launch
without having to know the attribute name which contains said command
line. When supplying --edit to the launch command an editor is opened
with the qemu command line inside. After closing the edit the qemu
command line is read from the file and written into the image.

Change-Id: Ifada25fe497ae04594ec85d4fef8baf5a0f36e5d