introspection tests: provide obj addr of cap to user
For each capability, the kernel object listing contains the address of
the object the capability points to. This address might deviate from
the actual object address due to dynamic casting. This is the case for
IPC gates without server permission. Therefore, the capabilities‘ object
addresses might be valuable additional information.

Change-Id: I559640ae8d17c3f6873983d1fbd450e4a2a93f8f