L4::TapWrapper - Also run kill_ps_tree during exit_test
In change Ib8198c50e872c49a5318ab3629ea6d817b065bb3 we moved the call to
kill_ps_tree from exit_test to finalize, to terminate qemu before the
processing through the external plugin is done. This caused a regression
where kill_ps_tree is not called at all if tap-wrapper does not call
finalize. This happens for example if tap-wrapper calls fail_test, which
has no need to call finalize as it produces its own TAP. With this
change we add the call the kill_ps_tree in exit_test again. We can do
this because kill_ps_tree will only kill children the first time it is called.

Change-Id: I17b9077015b45d9e9d7edc4b879d9c979785b045