L4::TapWrapper::Util - Get process tree via procfs
Up until now we either had to have an extra dependency or rely on
pstree to clean up all subprocesses. But we don't want  neither a
further dependency nor rely on a tool which doesn't give the necessary
information in a machine-readable format. To avoid both, we add code
that retrieves the relevant information directly from the procfs, as
the tools we did rely on would also do.

Change-Id: Ieb262b0417f20118c20edb34e404f3b6f82d92a5