Allow BID to collect diagnostics
This commit adds a feature where, if the environment variable
BID_COLLECT_DIAGNOSTICS is set, the buildsystem will create .diag files
for every object file where a compiler diagnostics is given. The file is
placed next to the corresponding object file.

It also adds targets to prog.mk and the main Makefile to list the
diagnostics encountered for that target (or all targets in an object
directory).

The diagnostics can, depending on compiler support, be formated as text,
the OASIS standardized SARIF format (Static Analysis Results
Interchange Format) or JSON.

If BID_COLLECT_DIAGNOSTICS is set to json then (for GCC 9 and newer) the
diagnostics are given in JSON format.

If BID_COLLECT_DIAGNOSTICS is set to sarif then (for GCC 13 and newer
and for clang 16 and newer) the diagnostics will be given in SARIF
format.

For any other value the dignostics will be given in Textformat.

Since this is done on a per-compiler-invocation basis diagnostics should
not be garbled, even by higly parallel builds.

Change-Id: Ifa098c5ede149802dd6cd918dadaced78cb24583