L4::TapWrapper::Plugin::BundleMode - add Nil UUID as Test-uuid
We want to have a uuid on every ok line, so we can easily find tests
where ok lines are missing. However there are a handful of cases where a
proper UUID does not make sense, because the underlying test doesn't
really test any functionality and just guard from some rare edge cases
going unnoticed, like this ok line that signals the complete result in a
bundle test. For these cases we want to use a uuid that satisfies
tooling, but can easily be distinguished from real uuids by a developer.
For this we use the Nil UUID which is a valid uuid consisting of all
zeroes.

Change-Id: I63f95aca5c84d2ce28e93b7d48152e0767adafa4