mk: support statically allocated heap on no-MMU systems
It is usually a bad idea to dynamically allocate RAM on no-MMU systems.
Hence the heap is directly reserved in the executable. Add the
infrastructure to pass the heap size to the linker script as special
symbol. Platforms still need to add the necessary boiler plate code to
their linker scripts together with a suitable libc backend.

Change-Id: Ia96a34683f684ab8d185d3f6cf15c742141b9a81