libgcc: Fix linking with ld.lld the LLVM linker
Previously the -PC$(COMPILER_RT) link dependency was added to LDFLAGS,
i.e. applied to both static and shared libraries. The -PC option,
however, is not directly intended for the linker, but for l4-bender.
But l4-bender is only used for shared libraries (.so), whereas for
incrementally linked object files (o.a) the linker is invoked directly.

The GNU linker seems to have ignored the unknown -PC option, not so
ld.lld which complains about it. Therefore -PC is now only used for
shared libraries.

Change-Id: I381e7440cc94feeb863665e9eee964f7d64459f9