mk: Do not inherit old deps when updating Kconfig.generated
Kconfig.generated keeps track of it dependencies at the time it was
generated, to detect if any of these dependencies has been removed,
in which case the Kconfig.generated file needs to be regenerated.

The recipe uses $^ when writing that dependency tracking file. That is
wrong, it cannot use $^ inside the $(KCONFIG_FILE) recipe to refer to
its dependencies, as this would include obsolete dependencies stored
in $(KCONFIG_FILE_DEPS) from the previous generation.

Introduce a variable that explicitly lists all the current dependencies
and use that instead of $^, so that no old dependencies are carried
over.

The same applies to the Kconfig.generated.platform(s,_type,.list) files.

Change-Id: I47d0d4ea3c937f5c1ca800e6334560fbf0fc355f