Support for externally overriding defconfig options
If the DROPSCONF_DEFCONFIG_OVERLAY variable is defined, it is
considered a path to a file whose contents are appended to the initial
configuration derived from a defconfig.

This enables to externally override any configuration option without the
need to alter the defconfig files in the source tree. For example, the
CONFIG_BID_USE_TOOLCHAIN_LIBGCC could be easily overriden like this.
This is useful since the state of this option is important during the
build directory creation already.

Change-Id: I19d8f67243bb022a8145753a91df4126ce055d58