Make sure the compiler type (Clang/GCC) is always detected
If $(CC) is set (either by the environment or by architecture), we
detect the compiler type (Clang/GCC) by running it directly. GCC is
always considered as the fallback alternative.

However, if $(CC) is not set explicitly by the environment and the
enviroment does not explicitly request Clang by setting $(CLANG), we
cannot run the compiler using $(callcc) before the $(BUILD_ARCH)
variable is defined (the value of $(CC) is derived from
$(CC_$(BUILD_ARCH)) in such cases).

Since we use $(BID_COMPILER_TYPE_f) in the $(KCONFIG_FILE).defines
target in the top-level makefile where $(BUILD_ARCH) is not defined yet,
the empty $(CC) leads to a strange output line during the build
directory setup phase in the case of using GCC:

  bash: line 1: --version: command not found

To avoid this, we use the $(CLANG) variable statically to identify
Clang and assume GCC otherwise when $(CC) is empty.

Change-Id: I960527927bcac8f413dd6e72eaf3dba8dd14d485