introspection testing: Include headers in kernel objects
A helper function for decoding kernel object printouts exists, but it
was not used by the introspection testing plugin, since it just ignored
the headers of the printout.

This change includes the headers in the decoded printout, so that the
helper function can be used. By this, we reduce code duplication in
upcoming plugins.

Change-Id: If465e1656241385f6d834c5f8a802cbfa418db79