Bail out with an error if a non-existent defconfig is requested
In case a specific default configuration is requested using the T
argument, but no such defconfig exists in mk/defconfig/, bail out with
an error message.

This is arguably a more sensible behavior than silently proceeding with
a different default configuration. Note that if T is not specified, this
does not change anything.

Change-Id: Ibd7af94be27841412ea83d2d8982344d0f010cd3