tool/bin/genexportpack - Enforce absolute target dir
When calling make exportpack with a relative path as target dir the
build system unexpectedly interprets the relative path to be relative to
the source tree instead of whereever make was called from. Since we're
unable to determine the original location because Make inherently
changes PWD, we resort to simply showing an explanatory error message
instead.

Change-Id: I713ab11bf51f88458543825d157068af457d04c9