Feature to specify a config template directory
The build system only allows to specify default configuration templates
from the 'mk/defconfig' directory for creating new build directories or
for checkbuild targets.

For specific customization of builds it would be handy to allow use-case
specific configuration templates to be used. Customer builds could be
customized for example.

This change introduces two new variables 'TEMPLDIR' and 'DFL_TEMPLATE'
which can be overriden during the make call to point the build system to
an alternative directory to pick config templates from. The feature
works analog to the kernel build system feature.

Change-Id: I00a80de374de64dff5a139626efb41dd16b59d6a