L4::TapWrapper - Do not compensate empty TAP
If we never get any tap lines nor a tap plan, TapWrapper until now
compensated this by adding a "1..0" tap plan. This is a problem, because
we might not be able to find faulty (esp. missing) tap output. To fix
this only ever add a tap plan at the end if at least one tap plan was
found among the gathered tap lines.

Change-Id: I1703642f2db0bad4109c27329ff2b097d14b9e25