Actually check directory parameters
Currently we only check whether a parameter was provided and not
whether an actual directory was specified. This leads to a lot of
random error messages und the user has to gues the reason for the
error.

Here we replace the simple string check by an actual check for the
presence of the specified directory.

Change-Id: I8777ade73442e7c9c31f962dff7c54592d9df35c