mk/test.mk: Add test bundle mode for fewer reboots
This change adds an additional setting that can be specified for the
test.mk role. If TEST_BUNDLE_MODE is set to y, then all tests from that
Makefile will be aggregated into a single .t file and sequentially
executed as a test bundle.

The feature must first be globally enabled through the config setting
BID_ALLOW_TEST_BUNDLING=y.

This significantly reduces the number of required reboots.

Beware that changing this in the config will *not* remove old .t files,
leading to potential double test execution. Be sure to remove them
manually.

Change-Id: I49cd79d1da381b223dc767a1f1a54ce3607af8ca
Signed-off-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>