mk: make .d file deletion robust for many files
Directories and packages with many files to compile might exceed the
maximum allowed command line length. Thus use a temporary file to
specify the files to delete whenever there is a risk of execessive file
parameters.

This also means that the creation of the directory containing
GENERAL_D_LOC needs to be moved out from the rule, because variables and
functions in the whole recipe are expanded before any shell is invoked.
An order-only prerequisite is sufficient because changes to the
directory do not necessitate updates of GENERAL_D_LOC.

Change-Id: I2880577f2e639be705c760e0fa75179a02ba3ec8