Fix check{cc,cxx,cxx_nowarn} for talkative compilers
Depending on the setup, a compiler might print something whether or not
the flag is actually supported. Instead of relying on the output we
might want to rely on the exit code.

There's one special case: clang doesn't by default throw an error if it
does not know a -W<warning> flag. We can solve this by also adding
-Wunknown-warning-option=error. We can only add this for clang, because
gcc doesn't know about this warning and will throw an error.

Change-Id: I0b7b4aa788ded657eb4b0b04a6cda7457bf10f5f