config.inc - Fix kconfig rebuild on clean targets
The rebuild of kconfig on clean targets was triggered because
DROPSCONF_CONFIG aka auto.conf was deemed outdated. The reason for this
is that its dependencies contained directories. Directories are
essentially always deemed newer. We fix this by demoting the directory
dependencies from "normal-prerequisites" to the lesser known type of
order-only-prerequisites which are not checked for updates but only for
existence, and thus are perfect in this case. This also allows us to get
rid of the MKDIR for include/l4 inside kconfig_call and add it to the
DIRS_FOR_CONFIG variable.

The downside is that with kconfig only being built if necessary, the
fixdep tool also is not always present. To counter this we add a
genfixdep target which calls the kconfig Makefile to build
scripts/basic/*

We add a variable ROOT_MAKEFILE so Makeconf can decide whether it's been
included from a subdirectory or the main Makefile. This can be used to
build fixdep on demand, but only from the main Makefile. We keep the old
message for Make calls in subdirectories.

Change-Id: I9d57686612a957322a3f8f85fbdb09d1efc5fee2