mk/config.inc - Set correct KCONFIG_OBJ_DIR
Although originally it seems to be intended that kconfig builds its
object files and binaries into the tool/kconfig/ directory, it's
currently using scripts/kconfig/ which presumably is a remnant from
Linux. There does not seem to be a way to force kconfig to build in
tool/kconfig/ instead, without altering kconfig itself. Patching kconfig
itself would most likely just be overriden during the next kconfig
update. So we're just setting the KCONFIG_OBJ_DIR to the path kconfig is
currently using.

Change-Id: I5dac5e99204e579c719fb0ec8920eeb423d5cd6d