mk/config.inc - Also remove kconfig binaries on cleanall
We need a way to cleanup mconf and friends. When the build environment
changes, e.g. system updates or when moving a build directory to another
host altogether, and the original libraries mconf was built with do not
exist anymore it will not be rebuilt automatically, but throws errors
and exits.

Change-Id: I48d19c64e28522e1a1d0397eafa03de6d50d6f85