Introduce mrproper target
One of the recipes of the cleanall target involves the removal of
.config.all which is a problem because other recipes of the cleanall
target still rely on that file to be present. To fix this we move the
removal of the config files to a different target called mrproper, which
depends on cleanall. This way we ensure that the config is deleted last.

Requires-Change-Id: I6b67a048df07584f91cfbdd5b453a51b691a16c2
Change-Id: I735c41bb2cd4a2432c6616cfdd5cf7a754888c0a