mk/sandbox: fix builddir creation case
The case where there is no build directory yet but it is being create
through the B= mechanism needs a bit of special handling.

First, the part of the Makefile that enters the sandbox never needs a
configured build directory. We only require Makeconf for other aspects
of the system. We must force the buildsystem to ignore that the build
directory is not yet configured to be able to actually configure it
using the olddefconfig rule and its friends. Otherwise Makeconf will
fail us.

Second, the actual creation of the builddir, using B= needs to be done
outside of the sandbox, since we cannot mount the directory otherwise.
Afterwards, the buildsystem re-invokes itsself using olddefconfig as a
target and enters the sandbox. Here it is not yet possible that make
changes to the newly created object directory yet, since it does not
contain a Makefile. This is only created by the olddefconfig rule. So
instead, always change to the PWD (this is needed because the sandbox
starts in the root directory) and set the object directory using O= to
behave more akin to non-sandboxed execution.

Change-Id: I66fbb9b09d321996cc5b13eb213aa3955b4c1a51