grub2efiiso: Target for generating ISO images booting via UEFI
The grub2efiiso target generates an ISO image using GRUB2 (similarily to
the grub2iso target), but the boot loader actually uses the combined
UEFI image and the UEFI boot mechanism (instead of the multiboot2
mechanism) to boot the operating system.

This is useful when the UEFI functionality is required, but booting from
an ISO image (or a physical CD/DVD) is the only viable option.

Change-Id: Iba9c938e007bf881d51fb98c6b1ea7facf770650