tap-wrapper - Add separate channel for debug output for plugins
Plugins only have one channel to send information to the outside world,
namely TAP, which works for simple cases where debug information is
limited. However sometimes these plugins want to output vast amounts of
debug information as well, which enlarges the TAP output unnecessarily.
For these cases we introduce the add_log_line function to
L4::TapWrapper::Plugin. Lines passed to this function are forwarded
either to STDOUT or to a separate log called "plugin.log" depending on
whether prove is used or not.

Change-Id: I1bc15599b53ed22e80607fbb116d361e2be461d3