run_test: Add support for on_finished action in kunit tests
We've added a way to trigger a shutdown instead of a reboot after the
kunit test has run within Fiasco. This change facilitates using this
feature upon starting a kunit t-file. For this we introduce the
environment variable TEST_KUNIT_ON_FINISHED which takes one of the
option "none", "reboot" or "shutdown". It defaults to "none". Also any
other value is treated as "none".

For backwards compatibility reasons, run_test still checks the formerly
used environment variable TEST_KUNIT_RESTART if TEST_KUNIT_ON_FINISHED
is not used.

Requires-Change-Id: I58ff4c0cdc74a299c0f0f28348a6062f30a3e6d3
Change-Id: Iea4197508cfb9cea9b68314596b95be1917d9a0b