tool/imagebuilder/fvp - Add environment variables for coverage dumping
In order to get coverage data out of the ARM FVP simulator in a timely
manner, we need it to shutdown and dump a memory buffer which was
written by Fiasco.

This change adds code that reacts to the environment variables
COVERAGE_DUMP_FILE, COVERAGE_MEMBUF_START and COVERAGE_MEMBUF_SIZE, adds
the necessary fvp options accordingly.

Change-Id: Ib754c7060e801b970ff6262efcc1e9b738fb385f