Plugin::LLVMCov - Add support for coverage data dumped from memory
In order to get coverage data out of the ARM FVP simulator in a timely
manner Fiasco writes the coverage data into a memory buffer, which then
is dumped into a file by the ARM FVP simulator on shutdown. The LLVMCov
Plugin then needs to read that memory dump and process it.

This change adds code that sets the target path for the memory dump and
reads it later once the ARM FVP simulator has terminated.

You enable it by setting COVERAGE_MEMBUF to 1, and set
COVERAGE_MEMBUF_START and COVERAGE_MEMBUF_SIZE according to the settings
specified in the kernel configuration.

Change-Id: Ie2e67da93f35963b35ab81f6da4e60f3cc58f4e0