tap-wrapper - Delete plugin tmpdir on exit if requested
When using a separate TEST_PLUGIN_TMPDIR, the TapWrapper::plugintmpdir
is not cleaned up, as it would be if plugintmpdir resided within
testworkdir created by mktemp. This can quickly create a problem,
especially when running a lot of t-files.

To counter this we add code to tap-wrapper that cleans up plugintmpdir
explicitly if the environment variable TEST_PLUGIN_TMPDIR_CLEANUP is set
to 1.

Change-Id: I3e2a84dcad0d5b504d91bb1cf014f5b71f6d39b3