Grub: Enable video mode
Enable video mode for multiboot2 variant, getting rid of the
"no suitable video mode found" message by Grub, which confuses users as
the last message they see. With the recently added video mode support
for multiboot2 and enabled framebuffer support in the kernel, messages
of the system will be shown.

Change-Id: Ia48f12d7ad39f88fe53d2a72f172fdbdc6f7a0b0