pkgdeps: Add option to add variant as modification
It is now possible to add a variant as a modification to a
dependency subtree.

The `-M <variant> <nodes ...> -ME` option can be provided to the
pkgdep utility to add `<variant>` as an additional variant to all
nodes of the subtrees of `<nodes...>`.
This way, all nodes keep the variants they need, but also get those,
that are added as modification on top.

Change-Id: I079ee2f2f4804fb6c6da0d2fb10d4da911e39473