Add Option to for directory specific variants
This Change adds the option to add a path to a variant in the
Control file to limit the variant to. That is done by putting the
path in square brackets after the variant.

The uvmm Control file has the following line:

```
variants: nofpu[server/src]
```

That leads to the whole package being build in `std` variant except
what's in `server/src`. That will be build in `nofpu` variant.

Change-Id: Ia840e5195fa90a1c19fe305922b475587dc07f47