mk: Automatically check requires / REQUIRES_LIBS consistency
By recording all the requires, provides and optional statements when
parsing the control file in a Make-readable fashion we are able to check
if all dependencies specified in REQUIRES_LIBS are actually specified in
the Control file (either as directly provided by the package itself, or
as a build-order dependency).

When an inconsistency is found, the build fails with an error message.

This should hopefully avoid inconsistencies in the future.

Change-Id: I8ef7bcf919eae087fa35fbbcfb4fdec7d5e9710d
Requires-Change-Id: Id588f7674a490e73d63e946e9b93e090e3363554
Requires-Change-Id: I125d4f037904da38e95f7245b35601b9043d123a