mk: Support advanced include file specifications
This change adds support for the syntax TARGET<SRC for include
filenames. Each filename that contains the < character is used such that
the file of the name SRC is installed under the name TARGET. This
includes potential directory parts of the SRC name, which are also
replaced.

For example

foo.h<contrib/foo_linux.h

will create a symlink pointing to contrib/foo.h in the include/
directory of the package. The symlink will be called foo.h. This can be
used to installed platform specific header files under their real
install name.

Change-Id: Ic5a5ae8953e2003fac370fe1200a2ce07bd78254