Makefile: Move .PHONY information to each target
Each target should have a phony code line and other phony lines
should not exist. The phony line then is removed with the target and
we avoid dangling phony targets. We also no longer have a bunch of
random phony lines every now and then.

This also removes the phonyness of compile_commands.json, since that
is just a file.

Change-Id: If42734902596ab1b69357b1bbe7f4fda4d0d6a6b