Put generic functionality into util.mk or Makeconf
This change adds a new file to the build system: util.mk. This file
provides a lot of rules, variables and functions that could be of
use at different points. util.mk has no includes itself and
therefore can be included at any point.

Some functionality is build system specific and therefore not in
util.mk, but in Makeconf.

Change-Id: I10b3e923b2e2b46c348f3501920fa5c1ca1d4b09