Add support for Ada / GNAT
This adds support for ADA libraries and programs built with gnat to the
buildsystem.

It also adds a feature to make CPP flags conditional, such
that they are only added for compile steps that support the C
preprocessor. The flags can be disabled by setting NO_CPP_foo, where foo
is the name of a compiler. In the concrete case this is disabled for Ada
by setting NO_CPP_ADAC=y.

Change-Id: I4b18574f7d0bf55fafe8962fe926c596a32abcff