Add SONAME support
Supporting an SONAME requires two things: letting the linker know the
intended SONAME and creating the symlinks to the actual file. L4Re is
supposed to follow the usual schema where the SONAME encodes just the
major version.

The shared object is stored with the full version suffix. Symlinks are
created for the plain .so file for linking and the major version for
runtime linking.

Change-Id: I5dde1f8ed15de7e44d61601c42d6219cb876a43a